gameModeRefactoring[REFA] : Reorganisation du dossier d'interfaces
