+++ src/app/shared/components/maps/map-departements-francais/data.ts
@@ -1,4 +1,4 @@
-import { Area } from '../../../utils/interfaces/area'
+import { Area } from '../../../utils/interfaces/map-oriented/area'
 
 let data : Area[] = [
     {
