+++ src/app/shared/utils/interfaces/game-oriented/game-mode-meta-data.interface.ts
