+++ src/app/shared/utils/interfaces/map-oriented/area.ts
