+++ src/app/shared/utils/interfaces/map-oriented/map-meta-data.interface.ts
