+++ src/app/shared/utils/interfaces/map-oriented/zone.ts
